removed code smells
